indicate which input line fails if parse failure
If any line of the input file is improperly formatted, the parse of that line will fail. This is hard to fix if you don't know which line fails, so I added a print stmnt indicating which line fails in case of parse failure (previously it just indicated the separator it expected to see).